Desktop: Share bundled runtime paths
